Add files via upload
A text compressor using Huffman coding is a software application designed to reduce the size of text documents by utilizing the Huffman coding algorithm. This algorithm assigns variable-length codes to different characters in a way that more frequently occurring characters are assigned shorter codes, resulting in an overall reduction in file size.